==> src/3. Co-expressionNetworkAnalysis.Rmd <==
---
title: "Mycobacterium Tuberculosis Project"
author: "Zaynab Mousavian"
date: "Sep. 2021"
output:
  pdf_document: default
  html_document:
    theme:
      bg: white
      fg: black
      base_font:
        google: Prompt
subtitle: Co-expression Network Analysis
---
  
**Outline:**  
- Loading all data and packages
- Sample clustering (Figure S2)
- Choosing the soft-thresholding power (Figure S4) 
- Co-expression network construction and clustering the network into modules  
- Merging of close modules  
- Quantifying module-trait associations 
- Visualizing all modules with PC1 vs. PC2 (Figure S6)

---
# 1. Loading all data and packages  
First of all, you need to load all necessary packages and also the data, prepared in the Preprocessing.Rmd file, before doing any further analysis.
```{r setup, include=TRUE,message=FALSE,warning=FALSE}
knitr::opts_chunk$set(echo = TRUE,message = FALSE,warning = FALSE)
percentage <- 0.3
corMeasure <- "pearson"
networkType <- "signed hybrid"
inputDirectory1 <- paste0("../Results/Preprocessing/",
                          as.character(percentage),"-")
inputDirectory2 <- paste0("../Results/DifferentialExpressionAnalysis/",
                          as.character(percentage),"-")
resultsDirectory <- paste0("../Results/Coexpression/",
                           as.character(percentage),"-",corMeasure,"-")

library(WGCNA)
library(flashClust)
library(ggplot2)
library(reshape)
library(ggrepel)
options(stringsAsFactors = FALSE)

lnames <- load(paste0(inputDirectory1,"ReadyData.RData"))
lnames
```

# 2. Sample clustering (Figure S2)
In this step, we detect which samples are outliers and maybe it would be better to remove those samples from the dataset.

```{r}
datExpr0 = as.data.frame(t(dat_Expr))
names(datExpr0) = row.names(dat_Expr)
rownames(datExpr0) = colnames(dat_Expr)

gsg = goodSamplesGenes(datExpr0, verbose = 0);
gsg$allOK

if (!gsg$allOK)
{
  # Optionally, print the gene and sample names that were removed:
  if (sum(!gsg$goodGenes)>0)
    printFlush(paste("Removing genes:", paste(names(datExpr0)[!gsg$goodGenes], 
                                              collapse = ", ")));
  if (sum(!gsg$goodSamples)>0)
    printFlush(paste("Removing samples:", paste(rownames(datExpr0)[!gsg$goodSamples],
                                              collapse = ", ")));
  # Remove the offending genes and samples from the data:
  datExpr0 = datExpr0[gsg$goodSamples, gsg$goodGenes]
}

row.names(datExpr0)=c(paste0("AS0",1:9),paste0("AS",10:20),paste0("LS0",1:9),paste0("LS",10:14),paste0("HS0",1:9),"HS10")
sampleTree = flashClust(dist(datExpr0), method = "average");
# Plot the sample tree: Open a graphic output window of size 12 by 9 inches
# The user should change the dimensions if the window is too large or too small.

pdf(file = paste0(resultsDirectory,"Figure S2.pdf"), width = 15, height = 9);
par(cex = 2);
par(mar = c(1,6,2,0))

datTraits=c(rep("#2F8AC4",20),rep("#E48725",14),rep("#A5AA99",10))
plotDendroAndColors(sampleTree,datTraits,groupLabels="",main = "Sample clustering to detect outliers", sub="", xlab="",
     cex.main=2,cex.dendroLabels = 1.5,cex.rowText = 2,face=2)

# Plot a line to show the cut
abline(h = 10, col = "red");

# Determine cluster under the line
clust = cutreeStatic(sampleTree, cutHeight = 10, minSize = 10)
table(clust)
# clust 1 contains the samples we want to keep.
keepSamples = (clust==1)
datExpr = datExpr0[keepSamples, ]
nGenes = ncol(datExpr)
nSamples = nrow(datExpr)
groupsNew=groups[keepSamples]

samplesOrder = c(row.names(datExpr[groupsNew=="Control",]),row.names(datExpr[groupsNew=="Latent",]),row.names(datExpr[groupsNew=="Active",]))
datTraits <- data.frame(Group=groupsNew)
rownames(datTraits) <- rownames(datExpr)
names(datTraits) <- "Group"
datTraits$group[datTraits$Group == 'Active']="2"
datTraits$group[datTraits$Group == 'Latent']="1"
datTraits$group[datTraits$Group == 'Control']="0"
datTraits$group=as.numeric(datTraits$group)
datTraits=datTraits[,-1,drop=FALSE]
names(datTraits) <- "Group"
datExpr = datExpr [samplesOrder, ]
datTraits = datTraits [samplesOrder, ,drop=FALSE]
# Re-cluster samples
sampleTree2 = flashClust(dist(datExpr), method = "average")
# Convert traits to a color representation: white means low, red means high, grey means missing entry
traitColors = numbers2colors(datTraits, signed = FALSE);

# Plot the sample dendrogram and the colors underneath.
plotDendroAndColors(sampleTree2, traitColors,
                    groupLabels = names(datTraits),
                    main = "Sample dendrogram and trait heatmap",cex.main=2)
abline(h = 10, col = "red");
dev.off()
save(datExpr,datTraits,file=paste0(resultsDirectory,"dataInput.RData"))
```

# 3.Choosing the soft-thresholding power (Figure S4)
Most of the biological networks have the scale-free topology. So, we try different powers factor in this step to choose the most suitable power factor which makes a network with the scale-free property.

```{r}
load(paste0(resultsDirectory,"dataInput.RData"))
# Choose a set of soft-thresholding powers
powers = c(c(1:10), seq(from = 12, to=20, by=2))

# Call the network topology analysis function
pearsonFunc<- function(){
  sft <<- pickSoftThreshold(datExpr, powerVector = powers, verbose = 0,
                            networkType=networkType)
}
bicorFunc<- function(){
  sft <<- pickSoftThreshold(datExpr, powerVector = powers, verbose = 0, corFnc="bicor",
                            networkType=networkType, corOptions = list(maxPOutliers =0.1))
}
ifelse(corMeasure=="bicor",bicorFunc(),pearsonFunc())

# Plot the results:
pdf(file=paste0(resultsDirectory,"Figure S4.pdf"),width=9,height=5)

par(mfrow = c(1,2));
cex1 = 0.5;
# Scale-free topology fit index as a function of the soft-thresholding power
plot(sft$fitIndices[,1], -sign(sft$fitIndices[,3])*sft$fitIndices[,2],
     xlab="Soft Threshold (power)",ylab="Scale Free Topology Model Fit,signed R^2",type="n",
     main = paste("Scale independence"),cex.main=0.7,cex.lab=1,cex.axis=0.7);
text(sft$fitIndices[,1], -sign(sft$fitIndices[,3])*sft$fitIndices[,2],
     labels=powers,cex=1,col="red");
# this line corresponds to using an R^2 cut-off of h
abline(h=0.90,col="red")
# Mean connectivity as a function of the soft-thresholding power
plot(sft$fitIndices[,1], sft$fitIndices[,5],
     xlab="Soft Threshold (power)",ylab="Mean Connectivity", type="n",
     main = paste("Mean connectivity"),cex.main=0.7,cex.lab=1,cex.axis=0.7)
text(sft$fitIndices[,1], sft$fitIndices[,5], labels=powers, cex=1,col="red")
dev.off()
```

# 4.Co-expression network construction and clustering the network into modules
We can construct a signed hybrid network using different correlation measures, like pearson and bicor, to see which network makes better sense in biology. In the signed hybrid network, only positive correlations among proteins expressions are included in the network and all negative correlations are removed.

```{r}
softPower = sft$powerEstimate;
softPower <- ifelse(is.na(softPower),4,softPower)
pearsonAdjancecy <- function(){
  adjacencyMat <<- adjacency(datExpr, power = softPower, type=networkType);
}
bicorAdjancecy <- function(){
adjacencyMat <<- adjacency(datExpr, power = softPower, corFnc="bicor", type=networkType
                           ,corOptions="maxPOutliers = 0.1");
}
ifelse(corMeasure=="bicor",bicorAdjancecy(),pearsonAdjancecy())

# Turn adjacency into topological overlap
TOM = TOMsimilarity(adjacencyMat,TOMType="signed");
dissTOM = 1-TOM

# Call the hierarchical clustering function
geneTree = flashClust(as.dist(dissTOM), method = "average");
# Plot the resulting clustering tree (dendrogram)
pdf(file=paste0(resultsDirectory,"Figure S2.pdf"),width=12,height=9)
plot(geneTree, xlab="", sub="", main = "Gene clustering on TOM-based dissimilarity",
     labels = FALSE, hang = 0.04);

# We like large modules, so we set the minimum module size relatively high:
minModuleSize = 5;
# Module identification using dynamic tree cut:
dynamicMods = cutreeDynamic(dendro = geneTree, distM = dissTOM,
                             deepSplit = 4, pamRespectsDendro = FALSE,
                            minClusterSize = minModuleSize);
table(dynamicMods)
# Convert numeric lables into colors
dynamicColors = labels2colors(dynamicMods)
table(dynamicColors)
# Plot the dendrogram and colors underneath
#sizeGrWindow(8,6)
plotDendroAndColors(geneTree, dynamicColors, "Dynamic Tree Cut",
                    dendroLabels = FALSE, hang = 0.03,
                    addGuide = TRUE, guideHang = 0.05,
                    main = "Gene dendrogram and module colors")
dev.off()
```

# 5. Merging of close modules
All modules are clustered according to the correlations between their module eigengenes and all modules which are close to each other, can be merged into only one module.

```{r}
# Calculate eigengenes
MEList = moduleEigengenes(datExpr, colors = dynamicColors)
MEs = MEList$eigengenes
# Calculate dissimilarity of module eigengenes
MEDiss = 1-cor(MEs);
# Cluster module eigengenes
METree = flashClust(as.dist(MEDiss), method = "average");
# Plot the result
pdf(file=paste0(resultsDirectory,"modulesClustering.pdf"),width=7,height=6)

plot(METree, main = "Clustering of module eigengenes",
     xlab = "", sub = "")
MEDissThres = 0.25
# Plot the cut line into the dendrogram
abline(h=MEDissThres, col = "red")
# Call an automatic merging function
merge = mergeCloseModules(datExpr, dynamicColors, cutHeight = MEDissThres, verbose = 0)
# The merged module colors
mergedColors = merge$colors;
# Eigengenes of the new merged modules:
mergedMEs = merge$newMEs;


MEList = moduleEigengenes(datExpr, colors = mergedColors)
MEs = MEList$eigengenes
# Calculate dissimilarity of module eigengenes
MEDiss = 1-cor(MEs);
# Cluster module eigengenes
METree = flashClust(as.dist(MEDiss), method = "average");
# Plot the result
plot(METree, main = "Clustering of module eigengenes",
     xlab = "", sub = "")

sizeGrWindow(12, 9)
plotDendroAndColors(geneTree, cbind(dynamicColors, mergedColors),
                    c("Dynamic Tree Cut", "Merged dynamic"),
                    dendroLabels = FALSE, hang = 0.03,
                    addGuide = TRUE, guideHang = 0.05)
# Rename to moduleColors
moduleColors = mergedColors
# Construct numerical labels corresponding to the colors
colorOrder = c("grey", standardColors(50));
moduleLabels = match(moduleColors, colorOrder)-1;
MEs = mergedMEs;
dev.off()
```

# 6. Quantifying module-trait associations
In this step, the amount of correlation between module eigengenes and the tarit of interest is computed for all modules. In this study, we define the trait vector as a vector including 0, 1 and 2 values. For healty control, latent and active TB samples, we consider 0, 1 and 2 in the trait vector respectively.

```{r}
# Define numbers of genes and samples
nGenes = ncol(datExpr);
nSamples = nrow(datExpr);
# Recalculate MEs with color labels
MEs0 = moduleEigengenes(datExpr, moduleColors)$eigengenes
MEs = orderMEs(MEs0)
moduleTraitCor = cor(MEs, datTraits, use = "p");
moduleTraitPvalue = corPvalueStudent(moduleTraitCor, nSamples);

pdf(file=paste0(resultsDirectory,"ModuleTraits.pdf"),width=3.5,height=5)

# Will display correlations and their p-values
textMatrix = paste(signif(moduleTraitCor, 2), "\n(",
                   signif(moduleTraitPvalue, 1), ")", sep = "");
dim(textMatrix) = dim(moduleTraitCor)
par(mar = c(6, 8.5, 3, 3));
# Display the correlation values within a heatmap plot
labeledHeatmap(Matrix = moduleTraitCor,
               xLabels = "TB Progression",
               yLabels = names(MEs),
               ySymbols = names(MEs),
               colorLabels = FALSE,
               colors = greenWhiteRed(50),
               textMatrix = textMatrix,
               setStdMargins = FALSE,
               cex.text = 0.8,
               cex.main = 0.7,
               cex.axis=0.7,
               cex.lab=0.9,
               zlim = c(-1,1),
               main = paste("Module-trait relationships"))
dev.off()

# Define variable group containing the group column of datTrait
Group = as.data.frame(datTraits$Group);
names(Group) = "Group"
# names (colors) of the modules
modNames = substring(names(MEs), 3)
geneModuleMembership = as.data.frame(cor(datExpr, MEs, use = "p"));
MMPvalue = as.data.frame(corPvalueStudent(as.matrix(geneModuleMembership), nSamples));
names(geneModuleMembership) = paste("MM", modNames, sep="");
names(MMPvalue) = paste("p.MM", modNames, sep="");
geneTraitSignificance = as.data.frame(cor(datExpr, Group, use = "p"));
GSPvalue = as.data.frame(corPvalueStudent(as.matrix(geneTraitSignificance), nSamples));
names(geneTraitSignificance) = paste("GS.", names(Group), sep="");
names(GSPvalue) = paste("p.GS.", names(Group), sep="");
probes = names(datExpr)
geneInfo0 = data.frame(Proteins = probes,
                       moduleColor = moduleColors,
                       geneTraitSignificance,
                       GSPvalue)
modOrder = order(-abs(cor(MEs, Group, use = "p")));
for (mod in 1:ncol(geneModuleMembership))
{
  oldNames = names(geneInfo0)
  geneInfo0 = data.frame(geneInfo0, geneModuleMembership[, modOrder[mod]],
                         MMPvalue[, modOrder[mod]]);
  names(geneInfo0) = c(oldNames, paste("MM.", modNames[modOrder[mod]], sep=""),
                       paste("p.MM.", modNames[modOrder[mod]], sep=""))
}
geneOrder = order(geneInfo0$moduleColor, -abs(geneInfo0$GS.Group));
geneInfo = geneInfo0[geneOrder, ]
write.csv(geneInfo, file = paste0(resultsDirectory,"geneInfo.csv"))
save(geneInfo,file=paste0(resultsDirectory,"geneInfo.RData"))

filterGenes = function(x)
{
  module = x
  columnMM = paste0("MM.",module)
  columnPMM = paste0("p.MM.",module)
  geneInfo_Module = geneInfo[which(geneInfo$moduleColor==module),]
  selectedGenes = intersect(which(geneInfo_Module[,columnMM]>0),
                            which(geneInfo_Module[,columnPMM]<0.05))
  row.names(geneInfo_Module[selectedGenes,])
}
modules = lapply(modNames,function(x) filterGenes(x))
save(modules,modNames,file=paste0(resultsDirectory,"modules.RData"))
```

# 7. Visualizing modules with PC1 vs. PC2 (Figure S6)
In this section, we perform principal component analysis (PCA) on the expression profile of proteins of each module and used the first and the second principal components (PC1 and PC2) to show differences in protein levels.

```{r}
load(file=paste0(resultsDirectory,"modules.RData"))
module="turquoise"
column = match(module, modNames);
moduleGenes = modules [[column]]
pca <- prcomp(t(dat_Expr[moduleGenes,]), scale=TRUE)
pca_x = as.data.frame(pca$x)
pca_x$protein=row.names(pca_x)
color_by_group <- function(pca_x, xaxis, yaxis) {
  p <- ggplot(pca_x, aes(y = .data[[yaxis]], x = .data[[xaxis]]))
  p + geom_point(aes(color=groups,size=4)) + theme_bw() +
    theme(legend.title = element_blank(), legend.position="none", axis.title = element_blank(),
          axis.text = element_blank(),panel.border = element_rect(colour = module, fill=NA, size=3)) + 
    scale_color_manual(values=c("#2F8AC4", "#E48725", "#A5AA99")) + 
    labs(x=xaxis, y=yaxis)+
    geom_text(label=pca_x$protein)
}
color_by_group(pca_x, "PC1", "PC2") 
pdf(paste0("../Results/Coexpression/",module,".pdf"))
color_by_group(pca_x, "PC1", "PC2") 
dev.off()

```
