==> src/5. SignatureRefinement.Rmd <==
---
title: "Mycobacterium Tuberculosis Project"
author: "Zaynab Mousavian"
date: "Oct. 2021"
output:
  pdf_document: default
  html_document:
    theme:
      bg: white
      fg: black
      base_font:
        google: Prompt
subtitle: SignatureRefinement
---
  
**Outline:**  
- Loading all data and packages  
- Lower respiratory infections datasets
- Sarcoidosis datasets
- Displaying all scatterplots (Figure 4) 

---
# 1. Loading all data and packages  
First of all, you need to load all necessary packages and the data prepared before to run differential expression analysis on each dataset.

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(limma)
library(ggplot2)
library(cowplot)
```

# 2. Lower respiratory infections datasets
We performed differential expression analysis using multiple transcriptomic datasets including either viral or
bacterial lower respiratory tract infections (LRTI) or systemic inflammatory response syndrome (SIRS).To this end, we used three transcriptomic datasets, GSE42026, GSE40012 and GSE60244, each containing various types of lower respiratory infections.
```{r}
studies = c("GSE42026","GSE60244","GSE40012")

studyName = studies[1]
load(file=paste0("../Datasets/",studyName,"/ReadyData.RData"))
control_Samples = row.names(data_Traits[data_Traits$Type=="Control",,drop=FALSE])
disease_Samples = row.names(data_Traits[data_Traits$Type!="Control",,drop=FALSE])
groups = data_Traits$Type
f <- factor(groups)
design <- model.matrix(~0+f)
colnames(design) <- levels(f)
cont.matrix <- makeContrasts(Bacterial-Control,H1N109-Control,RSV-Control,levels=design)
fit <- lmFit(data_Expr, design)
fit2 <- contrasts.fit(fit, cont.matrix)
fit.b <- eBayes(fit2)
save(fit.b,file=paste0("../Results/DifferentialExpressionAnalysis/",studyName,"-DE-Results.RData"))

studyName = studies[2]
load(file=paste0("../Datasets/",studyName,"/ReadyData.RData"))
control_Samples = row.names(data_Traits[data_Traits$Type=="Control",,drop=FALSE])
disease_Samples = row.names(data_Traits[data_Traits$Type!="Control",,drop=FALSE])
groups = data_Traits$Type
f <- factor(groups)
design <- model.matrix(~0+f)
colnames(design) <- levels(f)
cont.matrix <- makeContrasts(Bacteria-Control,Virus-Control,Coinfection-Control,levels=design)
fit <- lmFit(data_Expr, design)
fit2 <- contrasts.fit(fit, cont.matrix)
fit.b <- eBayes(fit2)
save(fit.b,file=paste0("../Results/DifferentialExpressionAnalysis/",studyName,"-DE-Results.RData"))


studyName = studies[3]
l=load(file=paste0("../Datasets/",studyName,"/ReadyData-Day0.RData"))
groups = data_Traits$Type
f <- factor(groups)
design <- model.matrix(~0+f)
colnames(design) <- levels(f)
cont.matrix <- makeContrasts(Bacterial-Control,SIR-Control,InfluenzaA-Control,Mixed-Control,levels=design)
fit <- lmFit(data_Expr, design)
fit2 <- contrasts.fit(fit, cont.matrix)
fit.b <- eBayes(fit2)
save(fit.b,file=paste0("../Results/DifferentialExpressionAnalysis/",studyName,"-DE-Results.RData"))

```

# 3. Sarcoidosis datasets
We also conducted the same experiment using three datasets from two studies containing sarcoidosis samples.

```{r}
load("../Datasets/curatedTBData/curatedTBData.RData")
studies=c("GSE83456","GSE42830","GSE42826")
DEG <- function(studyName){
  
  data_Expr = assay(object_list[[studyName]])
  pheno_Data = colData(object_list[[studyName]])
  data_Traits= as.data.frame(pheno_Data[,"TBStatus",drop=FALSE])
  names(data_Traits)="Type"
  
  data_Traits$Type=gsub("Control","Control",data_Traits$Type)
  data_Traits$Type=gsub("LTBI","Latent",data_Traits$Type)
  data_Traits$Type=gsub("PTB","Active",data_Traits$Type)
  data_Traits$Type=gsub("OD","Other",data_Traits$Type)
  
  control_Samples = row.names(data_Traits[data_Traits$Type=="Control",,drop=FALSE])
  disease_Samples = row.names(data_Traits[data_Traits$Type=="Other",,drop=FALSE])
  data_Expr=data_Expr[,c(control_Samples,disease_Samples)]
  data_Traits=data_Traits[c(control_Samples,disease_Samples),,drop=FALSE]
  
  if (studyName=="GSE42826"){
    cancer_Samples=c("GSM1050776","GSM1050795","GSM1050811","GSM1050814","GSM1050817","GSM1050818","GSM1050840","GSM1050855")
    peneumenia_Samples = c("GSM1050777","GSM1050801","GSM1050834","GSM1050857","GSM1050863","GSM1050830")
    data_Expr=data_Expr[,-which(colnames(data_Expr) %in% c(cancer_Samples,peneumenia_Samples))]
    data_Traits=data_Traits[-which(rownames(data_Traits) %in% c(cancer_Samples,peneumenia_Samples)),,drop=FALSE]
  
  }
  if (studyName=="GSE42830"){
    cancer_Samples=c("GSM1050929","GSM1050936","GSM1050958","GSM1050961","GSM1050967","GSM1050980","GSM1050993","GSM1051018")
    peneumenia_Samples = c("GSM1050945","GSM1050960","GSM1050963","GSM1050965","GSM1050974","GSM1050990","GSM1051005","GSM1051019")
    data_Expr=data_Expr[,-which(colnames(data_Expr) %in% c(cancer_Samples,peneumenia_Samples))]
    data_Traits=data_Traits[-which(rownames(data_Traits) %in% c(cancer_Samples,peneumenia_Samples)),,drop=FALSE]
  
  }
    
  groups = data_Traits$Type
  f <- factor(groups)
  design <- model.matrix(~0+f)
  colnames(design) <- levels(f)
  cont.matrix <- makeContrasts(Other-Control,levels=design)
  fit <- lmFit(data_Expr, design)
  fit2 <- contrasts.fit(fit, cont.matrix)
  fit.b <- eBayes(fit2)
  save(fit.b,file=paste0("../Results/DifferentialExpressionAnalysis/",studyName,"-DE-Results.RData"))

}
lapply(studies,function(x) DEG(x))
```

# 3. Visualizing results by scatterplots
This function generates a scatterplot to show the fold change values of proteins in our olink dataset (i.e. Active TB vs. control) compared to their fold changes in other non-TB datasets (i.e. non-TB diseases vs. control).

```{r}
DEPlot_func <- function(inputData, xaxis, yaxis,xpvalue, ypvalue,xlab,ylab) {
inputData$TBsig <- ifelse(inputData[,xpvalue] <= 0.05&abs(inputData[,xaxis]) > 1, inputData$TBsig <- "Sig", inputData$TBsig <- "nSig")
inputData$Othersig <- ifelse(inputData[,ypvalue] <= 0.05&abs(inputData[,yaxis]) > 1, inputData$Othersig <- "Sig", inputData$Othersig <- "nSig")
  ggplot(inputData, aes(x=.data[[xaxis]],y=.data[[yaxis]]))+
    geom_point(aes(color = Othersig,size=2)) +
    scale_x_continuous(limits=c(0,3)) +
    scale_y_continuous(limits=c(-0.5,2.5)) + 
    coord_fixed(ratio=1)  +
    scale_color_manual(values = c("grey", "red")) +
    theme_bw(base_size = 12) +
    geom_text_repel(
     data = subset(inputData, inputData[,ypvalue] <= 1 & abs(inputData[,yaxis]) > 0), 
      aes(label = Protein),
      size = 10,
      box.padding = unit(0.35, "lines"),
      point.padding = unit(2, "lines"),
      segment.colour = "transparent") +
    labs(y=paste0("Log2( ",ylab," )"), x=paste0("Log2( ",xlab," )"))+
  theme(legend.position = "none",axis.title = element_blank(),axis.text = element_text(size=16)) 
}
```

# 4. Displaying all scatterplots (Figure 4)
In this section, we used the DEPlot_func function defined above to display results as seen in Figure 4 of the manuscript.

```{r}
load("../Results/Coexpression/0.3-pearson-modules.RData")
load("../Results/DifferentialExpressionAnalysis/0.3-DE-Results.RData")
logFC <- as.data.frame(fit.b$coefficients)
pValue <- as.data.frame(fit.b$p.value)
DE_results <- cbind(logFC=logFC,pValue=pValue)
modules[[4]]=modules[[4]][-c(6,14)]#removing low-expressed proints in active TB (CSF-1 and LAP TGF-beta-1)
DE_results <- DE_results[modules[[4]],]
row.names(DE_results)=gsub("-","",row.names(DE_results))
row.names(DE_results)[3]="IFNG"
row.names(DE_results)[8]="CCL7"
row.names(DE_results)[13]="CD274"

load("../Results/DifferentialExpressionAnalysis/GSE42026-DE-Results.RData")
logFC <- as.data.frame(fit.b$coefficients)
pValue <- as.data.frame(fit.b$p.value)
new <- cbind(logFC=logFC,pValue=pValue)
new <- new[match(row.names(DE_results),row.names(new)),]
DE_results <- cbind(DE_results,new)

load("../Results/DifferentialExpressionAnalysis/GSE60244-DE-Results.RData")
logFC <- as.data.frame(fit.b$coefficients)
pValue <- as.data.frame(fit.b$p.value)
new <- cbind(logFC=logFC,pValue=pValue)
new <- new[match(row.names(DE_results),row.names(new)),]
DE_results <- cbind(DE_results,new)

load("../Results/DifferentialExpressionAnalysis/GSE40012-DE-Results.RData")
logFC <- as.data.frame(fit.b$coefficients)
pValue <- as.data.frame(fit.b$p.value)
new <- cbind(logFC=logFC,pValue=pValue)
new <- new[match(row.names(DE_results),row.names(new)),]
DE_results <- cbind(DE_results,new)

load("../Datasets/curatedTBData/curatedTBData.RData")
studies=c("GSE83456","GSE42830","GSE42826")
plotData <- function(studyName){
  load(paste0("../Results/DifferentialExpressionAnalysis/",studyName,"-DE-Results.RData"))
  logFC <- as.data.frame(fit.b$coefficients)
  pValue <- as.data.frame(fit.b$p.value)
  new <- cbind(logFC=logFC,pValue=pValue)
  new <- new[match(row.names(DE_results),row.names(new)),]
  names(new)=c(paste0("logFC.",studyName),paste0("pValue.",studyName))
  DE_results <<- cbind(DE_results,new)
}
lapply(studies, function(x) plotData(x))

DE_results$Protein = row.names(DE_results)
names(DE_results) <- lapply(names(DE_results), function(x) gsub(" - ",".",x))
names(DE_results)[19]="logFC.Bacterial2.Control"
names(DE_results)[23]="pValue.Bacterial2.Control"

pdf(file="../Results/ProteinSignature/Figure 4.pdf",width = 30,height = 50)
p1=DEPlot_func(DE_results,"logFC.Active.Control","logFC.Bacterial.Control","pValue.Active.Control","pValue.Bacterial.Control","Active vs. Control","Bacterial vs. Control")
p2=DEPlot_func(DE_results,"logFC.Active.Control","logFC.H1N109.Control","pValue.Active.Control","pValue.H1N109.Control","Active vs. Control","H1N109 vs. Control")
p3=DEPlot_func(DE_results,"logFC.Active.Control","logFC.RSV.Control","pValue.Active.Control","pValue.RSV.Control","Active vs. Control","RSV vs. Control")
p4=DEPlot_func(DE_results,"logFC.Active.Control","logFC.Bacterial.Control","pValue.Active.Control","pValue.Bacterial.Control","Active vs. Control","Bacterial vs. Control")
p5=DEPlot_func(DE_results,"logFC.Active.Control","logFC.InfluenzaA.Control","pValue.Active.Control","pValue.InfluenzaA.Control","Active vs. Control","InfluenzaA vs. Control")
p6=DEPlot_func(DE_results,"logFC.Active.Control","logFC.SIR.Control","pValue.Active.Control","pValue.SIR.Control","Active vs. Control","SIR vs. Control")
p7=DEPlot_func(DE_results,"logFC.Active.Control","logFC.Bacteria.Control","pValue.Active.Control","pValue.Bacteria.Control","Active vs. Control","Bacteria vs. Control")
p8=DEPlot_func(DE_results,"logFC.Active.Control","logFC.Virus.Control","pValue.Active.Control","pValue.Virus.Control","Active vs. Control","Virus vs. Control")
p9=DEPlot_func(DE_results,"logFC.Active.Control","logFC.Coinfection.Control","pValue.Active.Control","pValue.Coinfection.Control","Active vs. Control","Coinfection vs. Control")
p10=DEPlot_func(DE_results,"logFC.Active.Control","logFC.GSE42826","pValue.Active.Control","pValue.GSE42826","Active vs. Control","Sarcidosis vs. Control")
p11=DEPlot_func(DE_results,"logFC.Active.Control","logFC.GSE42830","pValue.Active.Control","pValue.GSE42830","Active vs. Control","Sarcoidosis vs. Control")
p12=DEPlot_func(DE_results,"logFC.Active.Control","logFC.GSE83456","pValue.Active.Control","pValue.GSE83456","Active vs. Control","Sarcoidosis vs. Control")

cowplot::plot_grid(p1,p2,p3,p4,p5,p6,p7,p8,p9,p10,p11,p12,ncol=3,nrow=4)
dev.off()
```
