==> src/8. SignaturePerformance.Rmd <==
---
title: "Mycobacterium Tuberculosis Project"
author: "Zaynab Mousavian"
date: "Nov. 2021"
output:
  pdf_document: default
  html_document:
    theme:
      bg: white
      fg: black
      base_font:
        google: Prompt
subtitle: SignaturePerformance
---
  
**Outline:**  
- Loading all data and R packages 
- Showing the performance of Signature (Figure 6B)
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# 1. Loading all data and R packages 
We firstly load all data and necessary R packages.

```{r setup, include=FALSE}
library(ComplexHeatmap)
library(circlize)
```

# 2. Showing the performance of Signature (Figure 6)
We stratified the individuals with active TB based on their PC1 values and a set of clinical features using a heatmap plot.

```{r}
file=read.csv("../Datasets/Preliminary Data/Olink heatmap.csv",sep=";")
names(file)=c("labID","Sex","CRP","ESR","Hb","Alb","Smear.positive","TB.manifestation","Cough","No.S.Sympt")
file=file[-12,]

l=load("../Results/Coexpression/0.3-pearson-modules.RData")
l=load("../Results/Preprocessing/0.3-ReadyData.RData")
dat_Expr=dat_Expr[,-c(3,29)]#Removing LAM006:P and LS0044:detected as outliers groups=groups[-c(3,29)]
batches=batches[-c(3,29)]
module = "turquoise"
column = match(module, modNames);
moduleGenes = modules [[column]]
moduleGenes = moduleGenes[-c(6,8,12,14)]
pca <- prcomp(t(dat_Expr[moduleGenes,]), scale=TRUE)
pca_x = as.data.frame(pca$x)
pca_x$protein=row.names(pca_x)


color_by_group <- function(pca_x, xaxis, yaxis) {
  p <- ggplot(pca_x, aes(y = .data[[yaxis]], x = .data[[xaxis]]))
  p + geom_point(aes(color=groups,size=4)) + theme_bw() +
    theme(legend.title = element_blank(), legend.position="none", axis.title = element_blank(),
          axis.text = element_blank(),panel.border = element_rect(colour = module, fill=NA, size=3)) + 
    scale_color_manual(values=c("#2F8AC4", "#E48725", "#A5AA99")) + 
    labs(x=xaxis, y=yaxis)+
    geom_text()
}
color_by_group(pca_x, "PC1", "PC2") 
pdf(paste0("../Results/Coexpression/Figure 6A.pdf"))
color_by_group(pca_x, "PC1", "PC2") 
dev.off()


PC1=pca_x[1:19,"PC1",drop=FALSE]
list=sort(PC1$PC1,index=TRUE)
PC1=PC1[list$ix,"PC1",drop=FALSE]
file=file[match(row.names(PC1),file$labID),]
file$PC1=PC1$PC1
file$Alb=as.numeric(file$Alb)

file=file[19:1,c("labID","TB.manifestation","Sex","PC1","CRP","SR","Hb","Alb","Smear.positive","Cough","No.B.symptoms")]
names(file)=c("labID","TB.manifestation","Sex","PC1","CRP","ESR","Hb","Alb","Smear.positive","Cough","No.S.sympt")

file[grep("pulm",file$TB.manifestation),"TB.manifestation"]="Pulmonary"
file$TB.manifestation=gsub("other ","Extrapulmonary",file$TB.manifestation)
file$TB.manifestation=gsub("lymphnode","Extrapulmonary",file$TB.manifestation)
file$TB.manifestation=gsub("disseminated","Disseminated",file$TB.manifestation)

file$Sex=gsub("F","Female",file$Sex)
file$Sex=gsub("M","Male",file$Sex)
file$Smear.positive=gsub("Y","Yes",file$Smear.positive)
file$Smear.positive=gsub("N","No",file$Smear.positive)
file$Cough=gsub("Y","Yes",file$Cough)
file$Cough=gsub("N","No",file$Cough)


colors=c(RColorBrewer::brewer.pal(8, "Set1"),RColorBrewer::brewer.pal(8, "Set2"),RColorBrewer::brewer.pal(8, "Dark2"),c("#2F8AC4", "#E48725", "#A5AA99", "#CD3A8E"),RColorBrewer::brewer.pal(8, "Pastel2"))
codes=c(RColorBrewer::brewer.pal(8, "Reds"),RColorBrewer::brewer.pal(8, "Greens"))
codes2=RColorBrewer::brewer.pal(8, "Dark2")

row.names(file)=file$labID
file=file[,-1]

ht_list=list()

col_struct_1 <- structure(c(codes2[1],codes2[3],codes2[6]),names=unique(file$TB.manifestation))
ht_list[[1]] <- Heatmap(file[,"TB.manifestation"],name="TB.manifestation",column_names_side  = "top",col = col_struct_1,width=unit(0.75,"cm"),height = unit(10,"cm"))

col <- colorRamp2(seq(min(file[,"CRP"],na.rm = TRUE), max(file[,"CRP"],na.rm = TRUE), length = 2), c("whitesmoke","#CD3A8E"))
ht_list[[2]] <- Heatmap(file[,"CRP"],na_col = colors[36],name="CRP",column_names_side  = "top",col=col,cluster_columns = FALSE,cluster_rows = FALSE,width=unit(0.75,"cm"),height = 0.25,heatmap_legend_param = list(legend_height = unit(2.5, "cm"),legend_width=unit(2,"mm")))

col <- colorRamp2(seq(min(file[,"ESR"],na.rm = TRUE), max(file[,"ESR"],na.rm = TRUE), length = 2), c("whitesmoke","#CD3A8E"))
ht_list[[3]] <- Heatmap(file[,"ESR"],na_col = colors[36],name="ESR",column_names_side  = "top",col=col,cluster_columns = FALSE,cluster_rows = FALSE,width=unit(0.75,"cm"),height = 0.25,heatmap_legend_param = list(legend_height = unit(2.5, "cm"),legend_width=unit(2,"mm")))

file[intersect(which(file$Sex=="Female"),which(file$Hb>120)),"Hb"]=130
file[intersect(which(file$Sex=="Male"),which(file$Hb>130)),"Hb"]=130
col <- colorRamp2(seq(min(file[,"Hb"],na.rm = TRUE), max(file[,"Hb"],na.rm = TRUE), length = 2), c("#CD3A8E","whitesmoke"))
ht_list[[4]] <- Heatmap(file[,"Hb"],na_col = colors[36],name="Hb",column_names_side  = "top",col=col,cluster_columns = FALSE,cluster_rows = FALSE,width=unit(0.75,"cm"),height = 0.25,heatmap_legend_param = list(legend_height = unit(2.5, "cm"),legend_width=unit(2,"mm")))

col <- colorRamp2(seq(min(file[,"Alb"],na.rm = TRUE), max(file[,"Alb"],na.rm = TRUE), length = 2), c("#CD3A8E","whitesmoke"))
ht_list[[5]] <- Heatmap(file[,"Alb"],na_col = "white",name="Alb",column_names_side  = "top",col=col,cluster_columns = FALSE,cluster_rows = FALSE,width=unit(0.75,"cm"),height = 0.25,heatmap_legend_param = list(legend_height = unit(2.5, "cm"),legend_width=unit(2,"mm")))

file["AS1002","Smear.positive"]=NA
file[which(file$Smear.positive=="unava"),"Smear.positive"]="Not applicable"
col_struct <- structure(c("#CD3A8E","whitesmoke",colors[36]),names=c("Yes","No","Not applicable"))
ht_list[[6]] <- Heatmap(file[,"Smear.positive"],na_col = colors[36],width=unit(0.75,"cm"),height = 0.25,name="Smear.positive",column_names_side  = "top",col = col_struct,cluster_rows = FALSE,cluster_columns = FALSE)

col_struct <- structure(c("#CD3A8E","whitesmoke"),names=c("Yes","No"))
ht_list[[7]] <- Heatmap(file[,"Cough"],na_col = colors[36],width=unit(0.75,"cm"),height = 0.25,name="Cough",column_names_side  = "top",col = col_struct,cluster_rows = FALSE,cluster_columns = FALSE)

col <- colorRamp2(seq(min(file[,"No.S.Sympt"],na.rm = TRUE), max(file[,"No.S.Sympt"],na.rm = TRUE), length = 2), c("whitesmoke","#CD3A8E"))
ht_list[[8]] <- Heatmap(file[,"No.S.Sympt"],na_col = colors[36],name="No.S.Sympt",column_names_side  = "top",col=col,cluster_columns = FALSE,cluster_rows = FALSE,width=unit(0.75,"cm"),height = 0.25,heatmap_legend_param = list(legend_height = unit(2.5, "cm"),legend_width=unit(2,"mm")))

col <- colorRamp2(seq(min(file[,"PC1"],na.rm = TRUE), max(file[,"PC1"],na.rm = TRUE), length = 2), c("#2F8AC4", "whitesmoke"))
ht_list[[9]] <- Heatmap(file[,"PC1"],na_col = colors[36],name="PC1",column_names_side  = "top",col=col,cluster_columns = FALSE,cluster_rows = FALSE,width=unit(0.75,"cm"),height = 0.25,heatmap_legend_param = list(legend_height = unit(2.5, "cm"),legend_width=unit(2,"mm")))

ht = ht_list[[1]] + ht_list[[2]]+ ht_list[[3]] + ht_list[[4]] + ht_list[[5]] + ht_list[[6]] + ht_list[[7]] + ht_list[[8]] + ht_list[[9]]

pdf(file="../Results/ProteinSignature/Figure 6B.pdf")
draw(ht, ht_gap = unit(c(2,rep(0,6),2), "mm"))
dev.off()
```
